backend/contexts/requirements.txt
Django==5.1.7
djangorestframework==3.16.0
django-cors-headers==4.7.0
dj-database-url==2.1.0
python-dotenv==1.1.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
gunicorn==23.0.0
pillow==11.2.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
typing_extensions==4.13.2
packaging==25.0
requests




